added some methods and solved the problems in -Wall -Wextra -Werror

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: bvelasco <bvelasco@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/01 13:45:22 by bvelasco          #+#    #+#             */
-/*   Updated: 2022/10/01 13:47:24 by bvelasco         ###   ########.fr       */
+/*   Updated: 2022/10/01 18:05:03 by bvelasco         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while ((s1[i] != 0 || s2[i] != 0) && i < n)
